requirements.txt
# I noticed a misbehaviour with 3.7.2 regarding cookie storage for raised HTTP response
# That's the case for the HTTPFound redirections
# It seems the cookies are not sent in the response, and therefore, the browser does not store them
# Obviously, sessions do not work then, and Openid Connect auth.py code is broken
aiohttp==3.8.1
aiohttp-cors==0.7.0
asyncpg~=0.24.0
pyyaml~=6.0
cryptography==39.0.1
jinja2~=3.0.2
#aiohttp_csrf
pymongo~=4.0.1
aiohttp-jinja2~=1.5
aiohttp-session~=2.9.0
aiohttp-middlewares==2.2.0
dataclasses-json~=0.5.6
StrEnum~=0.4.7 # Only until Python 3.11 is released (we can use the built-in StrEnum)
owlready2~=0.36
tqdm~=4.62.3
pydantic~=1.9.0
pyhumps
progressbar~=2.5
fastobo~=0.11.1
pronto~=2.4.4
requests==2.25.1
pandas==1.5.3
scipy==1.10.0
numpy==1.24.2
urllib3==1.26.13
#beautifulsoup4==4.11.2
#torch==1.11.0
obonet==0.3.1
networkx==2.8.8
python-decouple==3.8.0
pyjwt>=2.7.0
